# Use uintptr instead of runtime.Frame