Before:
```
BenchmarkErrorf/5-frames-4         	  158538	      7573 ns/op	    2992 B/op	      16 allocs/op
BenchmarkErrorf/10-frames-4        	   89179	     13124 ns/op	    4072 B/op	      26 allocs/op
BenchmarkErrorf/20-frames-4        	   44906	     26946 ns/op	    6232 B/op	      46 allocs/op
BenchmarkErrorf/40-frames-4        	   18634	     78400 ns/op	   13754 B/op	      87 allocs/op
```

After:

```
BenchmarkErrorf/5-frames-4         	  937490	      1248 ns/op	     384 B/op	       6 allocs/op
BenchmarkErrorf/10-frames-4        	  844726	      1510 ns/op	     384 B/op	       6 allocs/op
BenchmarkErrorf/20-frames-4        	  659119	      1842 ns/op	     384 B/op	       6 allocs/op
BenchmarkErrorf/40-frames-4        	  599695	      2035 ns/op	     384 B/op	       6 allocs/op
```